+++ package.json
@@ -119,7 +119,7 @@
     "@types/archiver": "^5.3.1",
     "@types/chrome": "^0.0.202",
     "@types/ember": "^4.0.3",
-    "@types/jest": "^29.2.3",
+    "@types/jest": "^29.5.12",
     "@types/jquery": "^3.5.14",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.17",
@@ -142,15 +142,15 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-webpack": "^0.13.2",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^27.1.5",
+    "eslint-plugin-jest": "^27.6.3",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-react": "^7.31.11",
     "fork-ts-checker-webpack-plugin": "^9.0.0",
     "glob": "^8.0.3",
     "husky": "^8.0.2",
-    "jest": "^29.3.1",
-    "jest-environment-jsdom": "^29.3.1",
-    "jest-localstorage-mock": "^2.4.22",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-localstorage-mock": "^2.4.26",
     "prettier": "^2.7.1",
     "pretty-quick": "^3.1.3",
     "request": "^2.88.2",
@@ -160,7 +160,7 @@
     "svg-url-loader": "^8.0.0",
     "terminate": "^2.5.0",
     "to-string-loader": "^1.2.0",
-    "ts-jest": "^29.0.3",
+    "ts-jest": "^29.1.2",
     "ts-loader": "^9.5.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.2",
