+++ src/extension/features/toolkit-reports/pages/outflow-over-time/utils.test.js
@@ -190,6 +190,7 @@ describe('Utils', () => {
       const excpected = [
         {
           name: 'Jan 2022',
+          type: 'line',
           data: [
             { x: 4, y: 100, custom: rawData },
             { x: 7, y: 600, custom: rawData },
@@ -200,6 +201,7 @@ describe('Utils', () => {
         },
         {
           name: 'Feb 2022',
+          type: 'line',
           data: [
             { x: 13, y: 200, custom: rawData },
             { x: 19, y: 900, custom: rawData },
