+++ src/extension/features/toolkit-reports/pages/root/components/report-filters/components/account-filter/container.ts
@@ -6,7 +6,7 @@ import { AccountFilterComponent } from './component';
 
 function mapContextToProps(context: ReportContextType) {
   return {
-    activeReportKey: context.selectedReport.key,
+    activeReportKey: context.selectedReport!.key,
   };
 }
