+++ src/extension/features/toolkit-reports/pages/root/components/report-selector/container.ts
@@ -6,7 +6,7 @@ import { ReportSelectorComponent } from './component';
 
 function mapContextToProps(context: ReportContextType) {
   return {
-    activeReportKey: context.selectedReport.key,
+    activeReportKey: context.selectedReport!.key,
     setActiveReportKey: context.setActiveReportKey,
   };
 }
